Fix up changes
