+++ tests/tools/test_analytics_handler.py
@@ -3,6 +3,8 @@ import pytest
 from src.agent.datasets.handlers.analytics_handler import (
     TREE_COVER_LOSS_ID,
     AnalyticsHandler,
+    _count_and_enrich,
+    format_id,
 )
 
 pytestmark = pytest.mark.asyncio(loop_scope="session")
@@ -95,3 +97,40 @@ async def test_build_payload_uses_no_canopy_cover_parameter():
         "forest_filter": None,
         "intersections": [],
     }
+
+
+# --- Nested (per-section) analytics results, e.g. Land GHG Inventory ----------
+# A nested result must NOT go through the flat aoi_id name-enrichment path, and
+# the data-point count comes from the first inner section's first list.
+
+
+async def test_nested_result_counts_and_skips_name_enrichment():
+    raw = {
+        "vegetation": {
+            "year": [2016, 2017, 2018],
+            "net_flux_MgCO2e": [1, 2, 3],
+        },
+        "agriculture": {
+            "category": ["cropland"],
+            "gross_emissions_MgCO2e": [9],
+        },
+    }
+    enriched, count = _count_and_enrich(
+        raw, [{"src_id": "BRA.25", "name": "São Paulo, Brazil"}]
+    )
+    assert count == 3  # first section's first list length
+    assert "name" not in enriched  # nested results skip aoi_id enrichment
+
+
+async def test_flat_result_enriches_names_and_counts():
+    aois = [
+        {"src_id": "BRA", "name": "Brazil"},
+        {"src_id": "ARG", "name": "Argentina"},
+    ]
+    raw = {
+        "aoi_id": [format_id("BRA"), format_id("ARG")],
+        "value": [10, 20],
+    }
+    enriched, count = _count_and_enrich(raw, aois)
+    assert count == 2
+    assert enriched["name"] == ["Brazil", "Argentina"]
